move config class and related tests to same folder
