--- src/__tests__/resolveTSConfig.test.ts
@@ -1,9 +1,0 @@
-import { resolveTSConfig } from '../cli';
-
-const regExForFilePath = /^\/([A-z0-9-_+]+\/)*(tsconfig+\.(json))$/gm;
-describe('resolveTSConfig', () => {
-    it('should resolve a tsconfig', async () => {
-        const tsconfig = await resolveTSConfig();
-        expect(tsconfig).toMatch(regExForFilePath);
-    });
-});
